#71 refactor done
