+++ src/main/java/de/yfu/intranet/methods/api/resources/mapper/MethodMapper.java
@@ -11,6 +11,8 @@ import org.mapstruct.Mapper;
 import de.yfu.intranet.methods.api.resources.MethodResource;
 import de.yfu.intranet.methods.data.domain.Method;
 
+import java.util.Set;
+
 @Mapper
 public interface MethodMapper {
 	
@@ -19,6 +21,9 @@ public interface MethodMapper {
 
 	Attachment mapToDataObject(AttachmentResource attachmentResource);
 	AttachmentResource mapFromDataObject(Attachment attachment);
+
+	Set<MethodType> mapToMethodTypes(Set<MethodTypeResource> methodTypeResources);
+	Set<MethodTypeResource> mapToMethodTypeResources(Set<MethodType> methodTypes);
 
 	MethodType mapToDataObject(MethodTypeResource methodTypeResource);
 	MethodTypeResource mapFromDataObject(MethodType methodType);
