--- src/main/java/de/yfu/intranet/methods/api/resources/mapper/SeminarMapper.java
@@ -1,14 +1,0 @@
-package de.yfu.intranet.methods.api.resources.mapper;
-
-import de.yfu.intranet.methods.api.resources.SeminarGoalResource;
-import de.yfu.intranet.methods.data.domain.SeminarGoal;
-import org.mapstruct.Mapper;
-
-@Mapper
-public interface SeminarMapper {
-
-    SeminarGoal mapToDataObject(SeminarGoalResource seminarGoalResource);
-    SeminarGoalResource mapFromDataObject(SeminarGoal seminarGoal);
-
-
-}
